New design (#1)
* something new

* fix position object in pair_provider.dart

* rewrite logic

* ui adjustment on multiple places

* further desing improvements

* Refactor updatePosition method in pair_provider.dart

* Refactor UI in AddLiquidityBox and RemoveLiquidityFormHook

* Refactor LiquidityInputField widget

* Refactor RemoveInput widget and RemoveLiquidityFormHook

This commit refactors the RemoveInput widget and RemoveLiquidityFormHook. It introduces a new RemoveInput widget that handles the UI for removing liquidity. The RemoveLiquidityFormHook class is also refactored to include new value notifiers for token and Zeniq errors. Additionally, the code now properly calculates the receive tokens based on the slider value.

* Refactor ManageButtons widget to use ToggleButtons

* Refactor ManageButtons widget to use LayoutBuilder for responsive width

* fix calculation of tokens in remove screen

* Refactor signing of transactions to use serializedUnsigned method